requirements.txt
streamlit
pandas
requests
beautifulsoup4
sqlalchemy
psycopg2-binary
plotly
python-dotenv